requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==5.0.0
pypdf2==3.0.1
python-docx==1.1.0
openai>=1.6.1,<2.0.0
tiktoken>=0.5.2,<0.6.0
numpy==1.24.3
pandas==2.0.3
plotly==5.17.0
python-dotenv==1.0.0 
huggingface-hub==0.33.4
transformers==4.53.2 